pid issue remains to be solved - revert to sth working
We've seen two instances of tapicero running on the production servers. The reason is that we are using two different places for the pid file - one for root and one for a normal user.

In this scenario tapicero started by root will not notice the tapicero run as a normal user and vice versa. One idea was to solve this by always using /var/run/pid. However this also requires write privileges to that place which essentially only root has.

So let's keep the two different places and just make sure we ALWAYS run tapicero as the same user.